Authentication

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,15 +1,23 @@
-import {useWorkoutContext} from "../hooks/useWorkoutContext.js";
+import { useWorkoutContext } from "../hooks/useWorkoutContext.js";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../hooks/useAuthContext.js";
 
-const WorkoutDetails = ({workout}) => {
-	const {dispatch} = useWorkoutContext();
+const WorkoutDetails = ({ workout }) => {
+	const { dispatch } = useWorkoutContext();
+	const { user } = useAuthContext();
 	const handleClick = async () => {
-		const response = await fetch(`/api/workouts/${workout._id}`, {
-			method: "DELETE"
+		if (!user) {
+			return;
+		}
+		const response = await fetch(`/api/workouts/${ workout._id }`, {
+			method: "DELETE",
+			headers: {
+				"Authorization": `Bearer ${ user.token }`
+			}
 		});
 		const data = await response.json();
 
-		if(response.ok){
+		if (response.ok) {
 			dispatch({ type: "DELETE_WORKOUT", payload: data });
 		}
 	};
@@ -17,10 +25,10 @@ const WorkoutDetails = ({workout}) => {
 	return (
 		<div className="workout-details">
 			<h4>{ workout.title }</h4>
-			<p><strong>Load (Kg): </strong>{workout.load}</p>
-			<p><strong>Reps: </strong>{workout.load}</p>
-			<p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true})}</p>
-			<span onClick={handleClick}>Delete</span>
+			<p><strong>Load (Kg): </strong>{ workout.load }</p>
+			<p><strong>Reps: </strong>{ workout.load }</p>
+			<p>{ formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true }) }</p>
+			<span onClick={ handleClick }>Delete</span>
 		</div>
 	);
 };
